+++ nextjs/nextjs-dynamic-routes/nextjs-dynamic-routes_lotr-app-dynamic-routes/.babelrc
@@ -1,0 +1,4 @@
+{
+    "presets": ["next/babel"],
+    "plugins": []
+  }
