+++ nextjs/nextjs-dynamic-routes/nextjs-dynamic-routes_lotr-app-dynamic-routes/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["next/core-web-vitals", "plugin:jest/recommended"],
+  "extends": ["next/babel", "next/core-web-vitals", "plugin:jest/recommended"],
   "env": {
     "jest": true
   },
