+++ nextjs/nextjs-dynamic-routes/nextjs-dynamic-routes_lotr-app-dynamic-routes/pages/volumes/index.js
@@ -1,26 +1,33 @@
 import Link from "next/link";
 import { introduction } from "../../lib/data";
+import { volumes } from "../../lib/data";
+import { useRouter } from "next/router";
 
 export default function Volumes() {
+  const router = useRouter()
+
+
+
+  function handleClick(){
+    const randomSlug = volumes[Math.floor(Math.random() * volumes.length)].slug
+    // const randomVolume = volumes[Math.floor(Math.random() * volumes.lenght)]
+    console.log(randomSlug);
+    router.push(`/volumes/${randomSlug}`)
+    
+  }
   return (
     <>
       <h1>The Lord of the Rings</h1>
       <p>{introduction}</p>
       <ul>
-        <li>
-          <Link href="/volumes/the-fellowship-of-the-ring">
-            The Fellowship of the Ring
-          </Link>
-        </li>
-        <li>
-          <Link href="/volumes/the-two-towers">The Two Towers</Link>
-        </li>
-        <li>
-          <Link href="/volumes/the-return-of-the-king">
-            The Return of the King
-          </Link>
-        </li>
+        {volumes.map(({ slug, title }) => 
+         (
+            <li key={slug}>
+              <Link href={`/volumes/${slug}`}>{title}</Link>
+            </li>
+          ))}
       </ul>
+      <button onClick={handleClick}>Go to a random Volume</button>
     </>
   );
 }
